% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simIOU.R
\name{makeSigma}
\alias{makeSigma}
\title{Make covariance matrix}
\usage{
makeSigma(beta, sigma, dt)
}
\arguments{
\item{beta}{Parameter beta of the OU process}

\item{sigma}{Parameter sigma of the OU process}

\item{dt}{Time interval}
}
\description{
Make covariance matrix
}
