% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPars.R
\name{plotPars}
\alias{plotPars}
\title{Plot posterior samples of movement parameters}
\usage{
plotPars(allpar, ind = NULL, log = FALSE)
}
\arguments{
\item{allpar}{Matrix of posterior samples}

\item{ind}{Vector of indices that should be included (e.g. to remove burn-in)}

\item{log}{Logical. If TRUE, plots posterior samples on log scale}
}
\description{
Plot posterior samples of movement parameters
}
