% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simIOU.R
\name{simIOU}
\alias{simIOU}
\title{Simulate from 2D IOU process}
\usage{
simIOU(obsTimes, beta, sigma, Q)
}
\arguments{
\item{obsTimes}{Vector of times of observations}

\item{beta}{Parameter beta of OU velocity process}

\item{sigma}{Parameter sigma of OU velocity process}

\item{Q}{Infinitesimal generator}
}
\value{
Simulated track
}
\description{
Simulate from 2D IOU process
}
