% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smooth_rcpp}
\alias{smooth_rcpp}
\title{Kalman filter and smoother}
\usage{
smooth_rcpp(data, param, Hmat, a0, P0)
}
\arguments{
\item{data}{Matrix of data, including (in that order) columns "x", "y", 
"time", and "state".}

\item{param}{Vector of movement parameters (beta and sigma)}

\item{Hmat}{Matrix of observation error variance (four columns, and one row 
for each row of data)}

\item{a0}{Initial state estimate vector}

\item{P0}{Initial estimate covariance matrix}
}
\value{
List of: log-likelihood, state estimates, and state estimate covariance
}
\description{
This code is adapted from the package crawl (Johnson et al., 2008).
}
\references{
Johnson, D. S., London, J. M., Lea, M. A., & Durban, J. W. (2008).
Continuous-time correlated random walk model for animal telemetry data
Ecology, 89(5), 1208-1215.
}
