% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateQ.R
\name{updateQ}
\alias{updateQ}
\title{Update transition rates}
\usage{
updateQ(nbStates, data, switch, priorShape, priorRate, priorCon)
}
\arguments{
\item{nbStates}{Number of states}

\item{data}{Matrix of all data (observations and transitions)}

\item{switch}{Matrix of transitions}

\item{priorShape}{Shape of prior gamma distribution for the transition rates}

\item{priorRate}{Rate of prior gamma distribution for the transition rates}

\item{priorCon}{Concentration of prior Dirichlet distribution for
transition probabilities}
}
\value{
Updated transition matrix
}
\description{
Update transition rates
}
